Consolidate rating validation
